+++ examples/example_apps/src/library_client/config/__init__.py
@@ -1,0 +1,1 @@
+"""Configuration owned by the importable client package."""